core: give the AI's exact sample rate to the timing and the frontend
The AI derives its rate by dividing the video clock by DACRATE, and the
quotient is very rarely a whole number: 48681812 / 2209 is 22037.94 Hz.
Two places rounded it down, in opposite directions.

get_dma_duration divided the transfer by a truncated samples-per-second,
which reported every DMA as up to 0.0043% too long.  Writing the
duration as LEN * (1 + DACRATE) / bytes_per_sample removes the rounding
entirely - the clock appears in both the counts-per-second and the
samples-per-second and cancels - and is simpler than what it replaces.

The rate handed to the frontend was the truncated integer, so a stream
arriving at 22037.94 Hz was declared as 22037 and the frontend resampled
at a permanently wrong ratio, 43 ppm off.  A frontend with dynamic rate
control absorbs that; one without accumulates it.  The backend
interface already carried the clock and the divider so the exact rate
could be reconstructed, and libretro takes the rate as a double.

Nothing was reading them, because main.c shadowed the backend's own
audio_out_backend_libretro with a file-static of the same name, built
from local extern declarations that had to be kept in step by hand and
were not: they still described set_frequency's old two-argument form,
so the backend's adapter was never installed and the clock and divider
were dropped.  Drop the shadow and use the interface the backend
defines.

The RSP audio path is untouched: the captured replay corpus and the
boot matrix are unchanged.

+++ libretro/libretro.c
@@ -148,7 +148,7 @@ uint32_t IgnoreTLBExceptions = 0;
 extern void     init_audio_libretro(void);
 extern void     deinit_audio_libretro(void);
 extern void     flush_audio_libretro(void);
-extern unsigned get_audio_sample_rate_libretro(void);
+extern double   get_audio_sample_rate_libretro(void);
 
 struct retro_rumble_interface rumble;
 
@@ -1087,7 +1087,7 @@ void retro_get_system_av_info(struct retro_system_av_info *info)
 
       info->timing.fps = fps;
    }
-   info->timing.sample_rate = (double)get_audio_sample_rate_libretro();
+   info->timing.sample_rate = get_audio_sample_rate_libretro();
 }
 
 unsigned retro_get_region (void)
